Добавляет методы для модели
