--- src/model/points.js
@@ -1,11 +1,0 @@
-import { getRandomPoint } from '../mock/point.js';
-
-const POINT_COUNT = 4;
-
-export default class PointsModel {
-  points = Array.from({length: POINT_COUNT}, getRandomPoint);
-
-  getPoints() {
-    return this.points;
-  }
-}
